% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die.R
\name{die}
\alias{die}
\alias{is.die}
\title{die}
\usage{
die(sides = 1:6, prob = rep(1/6, 6))

is.die(x)
}
\arguments{
\item{sides}{vector of die sides}

\item{prob}{vector of side probabilities}

\item{x}{an R object}
}
\value{
an object of class \code{"die"}
}
\description{
Creates an object of class \code{"die"}
}
\examples{
\dontrun{
# default
die1 <- die()}
\dontrun{
# another die
die2 <- die(c('a', 'b','c','d','e','f'))
}
is.die (die)
}
\seealso{
\code{\link{roll}}
}
