% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\alias{is.roll}
\title{roll a die}
\usage{
roll(x, times = 1)

is.roll(x)
}
\arguments{
\item{x}{an R object}

\item{times}{number of rolls}

\item{die}{object of class \code{"die"}}
}
\value{
an object of class \code{"roll"} with the following elements:

\item{rolls}{vector of rolls}

\item{sides}{vector of coin \code{"sides"}}

\item{prob}{vector of coin \code{"prob"}}

\item{total}{total number of tosses}

\item{1}{number of side 1} #{heads}{number of heads}

\item{2}{number of side 2}

\item{3}{number of side 3}

\item{4}{number of side 4}

\item{5}{number of side 5}

\item{6}{number of side 6}
}
\description{
Creates an object of class \code{"roll"}
}
\examples{
\dontrun{die1 <- die()

# roll a die 10 times
die10 <- roll(die1, times = 10)}
\dontrun{

# add 5 more rolls
roll15 <- roll10 + 5}
}
